다시보기가 중간에 엉뚱하게 죽던 것 — 이동 입력도 float32 로
dt 는 9-03 에 고쳤는데 이동 입력(move.x/y)이 남아 있었다. 저장은 float32
로 하면서 판은 float64 로 굴려서, 저장된 값이 실제로 쓴 값과 6e-8 쯤
어긋났다. 전기선 시뮬은 그보다 훨씬 예민해 그 정도로도 죽는 시점이
통째로 갈린다.

측정한 근거:
  · 정상 재현되는 기록 50개에 float32 오차만큼(6e-8) 넣어 봤더니
    3개(6%)가 다른 결말로 바뀌었다 — 실제 어긋나던 비율 3/53 과 같다.
  · 아날로그 입력 값 100만 개 중 float32 왕복에서 살아남는 비율
    0.0000% → fround 뒤 100%.
  · 키보드로 상하좌우만 누른 기록(-1/0/1)은 원래도 다 멀쩡했다.
    대각선(1/√2)과 터치 스틱처럼 어중간한 값에서만 어긋났다.

이미 올라간 기록은 되살릴 수 없다(원래 값이 없다). 그래서 재생을 켤 때
렌더 없이 한 번 미리 돌려 보고, 기록한 시간과 다르면 다시보기 바에
"실제 기록은 N초 · 옛 방식이라 재현이 어긋납니다" 라고 붙인다.

+++ public/js/input.js
@@ -153,8 +153,17 @@ export class Input {
       x /= len;
       y /= len;
     }
-    this.move.x = x;
-    this.move.y = y;
+    // dt 와 같은 이유로 float32 로 깎아서 내보낸다. 다시보기는 이 값을
+    // float32 로 저장하는데(encodeReplay), 판이 돌 때 float64 로 굴리면
+    // 저장된 값과 6e-8 쯤 어긋난다. 전기선 시뮬은 그보다 훨씬 예민해서,
+    // 그 정도 차이로도 죽는 시점이 통째로 갈린다 — 실제로 이 오차만큼
+    // 흔들어 보니 정상 기록 50개 중 3개(6%)가 다른 결말로 바뀌었고,
+    // 실제 어긋나던 비율(3/53)과 정확히 같았다.
+    //
+    // 키보드로 상하좌우만 누르면 -1/0/1 이라 float32 로도 정확해서 멀쩡했다.
+    // 대각선(1/√2)이나 터치 스틱처럼 어중간한 값일 때만 어긋났다.
+    this.move.x = Math.fround(x);
+    this.move.y = Math.fround(y);
 
     this.jumpPressed = this.jumpQueued;
     this.jumpQueued = false;
